# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IBrand": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICategory": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProduct": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"categoryId": {
						"type": "string"
					},
					"brandId": {
						"type": "string"
					},
					"stock": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"isPromotion": {
						"type": "boolean"
					},
					"brand": {
						"$ref": "#/components/schemas/IBrand"
					},
					"category": {
						"$ref": "#/components/schemas/ICategory"
					},
					"brandName": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"lowStockThreshold": {
						"type": "number",
						"format": "double"
					},
					"stockHistory": {
						"items": {
							"properties": {
								"quantity": {
									"type": "number",
									"format": "double"
								},
								"date": {
									"type": "string",
									"format": "date-time"
								}
							},
							"required": [
								"quantity",
								"date"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"price",
					"categoryId",
					"brandId",
					"stock",
					"isPromotion",
					"lowStockThreshold",
					"stockHistory"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CategoryCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ICategory_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"BrandCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IBrand_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/products/search": {
			"get": {
				"operationId": "SearchProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IProduct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "brand",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isPromotion",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "inStock",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/{productId}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IProduct"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IProduct"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN|ROLE_STORE_KEEPER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"categoryId": {
										"type": "string"
									},
									"brandId": {
										"type": "string"
									},
									"stock": {
										"type": "string"
									},
									"isPromotion": {
										"type": "string"
									},
									"lowStockThreshold": {
										"type": "string"
									},
									"image": {
										"type": "string",
										"format": "binary"
									},
									"existingImageUrl": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IProduct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProduct"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"categoryId": {
										"type": "string"
									},
									"brandId": {
										"type": "string"
									},
									"stock": {
										"type": "string"
									},
									"isPromotion": {
										"type": "string"
									},
									"lowStockThreshold": {
										"type": "string"
									},
									"image": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"name",
									"description",
									"price",
									"categoryId",
									"brandId",
									"stock",
									"isPromotion",
									"lowStockThreshold"
								]
							}
						}
					}
				}
			}
		},
		"/products/{productId}/stock": {
			"put": {
				"operationId": "UpdateStock",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IProduct"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN|ROLE_STORE_KEEPER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"quantity": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"quantity"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/products/{productId}/low-stock-threshold": {
			"put": {
				"operationId": "UpdateLowStockThreshold",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IProduct"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN|ROLE_STORE_KEEPER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"lowStockThreshold": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"lowStockThreshold"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/products/{productId}/stock-history": {
			"get": {
				"operationId": "GetStockHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"quantity": {
												"type": "number",
												"format": "double"
											},
											"date": {
												"type": "string",
												"format": "date-time"
											}
										},
										"required": [
											"quantity",
											"date"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN|ROLE_STORE_KEEPER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/{productId}/subscribe-restock": {
			"post": {
				"operationId": "SubscribeToProductRestock",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/{productId}/subscribe-price-change": {
			"post": {
				"operationId": "SubscribeToProductPriceChange",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/categories/{categoryId}/subscribe-new-products": {
			"post": {
				"operationId": "SubscribeToCategoryNewProducts",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/{productId}/unsubscribe-restock": {
			"delete": {
				"operationId": "UnsubscribeFromProductRestock",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/{productId}/unsubscribe-price-change": {
			"delete": {
				"operationId": "UnsubscribeFromProductPriceChange",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/categories/{categoryId}/unsubscribe-new-products": {
			"delete": {
				"operationId": "UnsubscribeFromCategoryNewProducts",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/{productId}/check-restock-subscription": {
			"get": {
				"operationId": "CheckProductRestockSubscription",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isSubscribed": {
											"type": "boolean"
										}
									},
									"required": [
										"isSubscribed"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/{productId}/check-price-change-subscription": {
			"get": {
				"operationId": "CheckProductPriceChangeSubscription",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isSubscribed": {
											"type": "boolean"
										}
									},
									"required": [
										"isSubscribed"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/categories/{categoryId}/check-new-products-subscription": {
			"get": {
				"operationId": "CheckCategoryNewProductsSubscription",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isSubscribed": {
											"type": "boolean"
										}
									},
									"required": [
										"isSubscribed"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICategory"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryCreationParams"
							}
						}
					}
				}
			}
		},
		"/categories/{categoryId}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ICategory"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ICategory"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ICategory_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/brands": {
			"get": {
				"operationId": "GetBrands",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IBrand"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBrand",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBrand"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BrandCreationParams"
							}
						}
					}
				}
			}
		},
		"/brands/{brandId}": {
			"get": {
				"operationId": "GetBrand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IBrand"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "brandId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateBrand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBrand"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "brandId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IBrand_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBrand",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": [
							"ROLE_ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "brandId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}